parser: support unquoted filenames with spaces in them
Which appears to be what git uses, as opposed to what I have mainly been
testing against using GNU diff.

This took a while to figure out as originally I did not realise that the
git style patches were placing a trailing tab after the filename in for
these patches. The tab seems to be the only way which we can possibly
distinguish these filenames from a timestamp since we also support using
spaces as a timestamp to filename separator.