parser: handle newlines in quoted filename strings
Found by playing around with filenames to see if adding certain
characters would cause git to quote the filename. Seems that it does!

This patch does seem to indicate that the current filename parser is
sorely missing a few different escape sequences though, as whatever
character a filesystem supports seems like it could theoretically end up
here.

Let's worry about that later for now and just fix this known case, and
try and figure out the rest of the escapes later.