parser: fix a path strip of zero in git extended headers
We were not correctly handling the special case of strip zero where the
'a/' and 'b/' prefixes are to be preserved. To implement this, since the
filename we are parsing is known to be a git style header, simply
prepend the a/ or b/ prefix to the filename to achieve the expected
behaviour.