patch: add support for rename patches that have already been renamed
If the file we are patching is the same as what the output file is going
to be - then the file to patch must have already been renamed. Detect
this case, and apply the patch as necessary.

When we detect this situation we still need to be applying the patch
content however. The file may have been renamed to the correct location
already, but the patches content may have not been applied yet.