parser: support parsing multiple zero content git extended patches
If there was no leading header in the patch then we would parse the
patch until the end of the file looking for it. While this does not
matter for a single git extended patch if there are mutliple of them
concatinated together this causes problems.

Handle this by detecting the situation where we have found a second git
diff header.