parser: add support for parsing patches with mode changes
Which has unfortunately shown a hole in the approach previously done for
parsing the filename out for rename patches. Unfortunately for mode
change only patches the filename only appears to be in the `diff --git`
header line. Add some very basic support for parsing this header so that
in the future we may also apply mode changes to patches.