patch: prompt to cerr when output patch to cout
Making more of the test_full.py tests pass when run against GNU patch.
The way that this is done is not very pretty. However functionally this
does allow for `-o -` to be used and for stdout from this command to be
redirected to somewhere else without user prompts getting in the way.