tests: do not pass through an empty commandline argument
This papers over a behavioural difference between this implementation of
patch and GNU patch. In our current implementation an empty commandline
string argument is treated as if it was never given in the first place.
On the other hand GNU patch makes this distinction. This was resulting
in this test failing for GNU patch, but not for our implementation.

Paper over this issue since it shouldn't really matter in reality, but
perhaps in the future we should also make this distinction.