cmdline: remove some uneeded error checking
We do not need to check whether the option is at the end as that is
already handled by the case that we are simply looking at the next
argument.

Furthermore there is no need to check whether we are already at the end
of the long option as currently we have no options which only contain
the long option.