parser: detect git binary patches
The git binary patch is actually simply a UTF16 file as I investigate
how to handle encodings across platforms!

This patch detects that we have found a git binary patch. In the future
I hope that we can parse this git binary patch properly and actually be
able to apply the patch. I suspect that we'll just go for the easy
approach for now and skip the patch like how GNU patch does.