patch: refuse to apply patches for non-regular files
Found by attempting to apply a patch which contained a git submodule
update. This requires the heuristic for determining which file to patch
to change as otherwise for such a patch we would otherwise consider it
impossible to determine the filepath for the git submodule which is
being patched.

If this scenario occurs, write to the reject file and proceed to the
next patch.

This requires a refactor of the applier as previously it was the only
place which would apply reject files. Now the driving patch loop may
also write to reject files.

There is some ugly duplication being added here that should be tidied up
in the future. Until the structure of apply patch is better defined,
leave the FIXME for now.