patch: add support for wide paths on windows for patch file path
On Windows convert the commandline to utf-8 before parsing it. Store the
patch filepath internally in the native format by passing it around as a
std::filesystem::path instead.

There may be better approaches here to supporting wide paths, but this
should hopefully serve as an initial baseline for future improvements to
come.