patch: add initial support for patching unicode paths on windows
The previously attempted approach of trying to use std::filesystem::path
everywhere that mattered was proving over complicated to implement.
Instead patch now internally handles every string as utf8 and at
filesystem boundries converts the utf8 path to its wide path equivalent.

Not every part of the code which needs this done is completed here.
However, a basic test case is now passing, so this should be a good
initial version.