patch: implement support for --read-only commandline option
Somehow I had missed that this option was a thing previously and thought
I was more feature complete with GNU patch than I was. Add support for
this flag, as well as some tests to go with it.

Also align a error message more closely with GNU patch while we are at
it.