patch: quote filenames containing shell characters
Filenames containing shell characters are now quoted. I had originally
thought that GNU patch was quoting non-ASCII chars, but testing using
the test_unicode_rename_git_no_quoting testcase, this assumption was
proven to be wrong. Reading the GNU manual, filenames instead appear to
be quoted if they contain shell metacharacters (whatever that means). In
my case, this was due to a '!'.