patch: fix atomicity of applying a git commit patch
As the comment in this commit explains a git commit is an atomic
collection of changes. Since a commit consists of multiple patches, we
must defer actually making any changes until the rest of the commit has
finished applying.

This commit implements an initial implementation of this. However, I
strongly suspect that this implementation is lacking in many regards. It
does make the previously failing test pass however!