cmdline: fix handling of multiple options in single position
We were not considering the situation that a short boolean option could
be followed by a string or integer argument in the same position. Fix
this by making string and integer handling more similar to how we
already handle booleans by iterating over every single option.

Unfortunately we still need to special case the reject format cases as
we convert those to an enum, but since they do not have short options it
should not cause any problems for this specific case for now.