tests: add some useful output when running tests
Output to stdout the start of the test, the success/fail/disablement of
that test, as well as the time that it takes to run that test.

Furthermore, when running the full test suite, say how many tests are
going to be run, and give a summary at the end with the time taken and
pass/fail.