tests: run each test with ctest individually
Similar to gtest_discover_tests, for each target that we want to apply
this to, run each test individually intead of the entire test binary.
This should somewhat improve discoverability with failed tests and some
better tooling.

The approach is somewhat naive right now, but for our purposes it should
make do for the moment.