parser: fix verbose output when parsing a patch with trailing garbage
Which also happens to nicely simplify the unified patch parsing
implementation! Turns out that all we need to do is after completely
parsing a hunk we can just look for the beginning of a new hunk header.
If we cannot happen to find it, then that must mean we are at the end of
a patch.