parser: improve error on a corrupted context patch
If a line being parsed is expected to contain a patch content line, but
the operation for that line is not valid then error out with a parser
error. Also ensure that we throw an informative error here that also
says the line number and the content of the line which was just parsed.

The stdout output here is the same as what is done for a unified patch,
but is not the same as what GNU patch does. Therefore, for now, we don't
add a test for it. Hopefully this behavioural difference isn't as a
result of some other underlying issue.

While we are at it, move all context patch specific tests into their own
file.