patch: fix being unable to add a file in a normal patch
Our approach of just using the filename of /dev/null to determine
whether we are adding a file or not does not work for normal patch as it
does not contain any patch header with a filename in it.

There is some code here that is starting to look pretty nasty, but I'm
not entirely sure yet of the behaviour which we intend to implement, so
that should be fine for now.