applier: fix --ifdef behaviour on line added and then removed
I was juts trying to add test coverage to this piece of code, and then
realised that the implementation was actually broken for this case! diff
does not seem to really produce diffs that have lines added before
removals, but support this anyway since it's still a valid patch file.