patch: implement options to enable/disable backups on mismatch
Add the options to enable or disable the behaviour of applying backups
on mismatch. The '--backup-if-mimatch' option is kind of useless at the
moment since it defaults to true, and the only way to turn it off is the
'--no-backup-if-mismatch' option. However, this should change as soon as
the '--posix' option is introduced which will be changing some of these
defaults!