parser: fix a case of incorrectly parsing normal diff ranges
I created a normal diff where I was trying to have the file removed.
It wasn't! The parsing of the patch was actually not working altogether!
Turns out that the root cause stemmed from the normal range not being
parsed as expected, and the parser was trying to fetch one two many
lines from the normal patch.