patch: fix git patches not correctly removing empty files
The deferred write mechanism was resulting in the following
remove-if-empty check to incorrectly determine that there was some
mismatch in applying the removal patch, so we should not be removing the
file altogether.

This _may_ be hacky (and I am not sure if this causes other problems
yet), but to fix this, we now immediately apply git removal patches
instead of deferring it.