patch: begin adding initial support for handling symlinks
git patches seem to have support for symlinks. From what I can tell,
this is signalled by the mode of the patch file.

For now, only add support for the "symbollic link" prompt which is given
when given a symlink patch. Soon, we should also handle actually adding
the symlink itself!