parser: factor some functions to make use of line parser
In particular, this has the nice side effect of slightly simplifying
parse_git_header_name to not have a dead check for the git header.

We are also able to remove LineParser::end() and LineParser::current()
with these changes.

Finally, this allows us to get rid of the substr with magic number
checks when parsing file names.